pyproject.toml
[tool.poetry]
name = "PPG-Log"
version = "0.1.0"
description = "Helper Tools for PPG FlySight Flight Log Management"
authors = ["sco1 <sco1.git@gmail.com>"]

readme = "README.md"
homepage = "https://github.com/sco1/"
repository = "https://github.com/sco1/ppg-log"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering",
    "Typing :: Typed",
]

[tool.poetry.dependencies]
python = "^3.10"
humanize = "^4.1"
kaleido = "^0.2, != 0.2.1.post1"
pandas = "^1.4"
peewee = "^3.14"
plotly = "^5.8"
python-dotenv = "^0.20"

[tool.poetry.dev-dependencies]
black = "^22.3"
bump2version = "^1.0"
flake8 = "^4.0"
flake8-annotations = "^2.8"
flake8-bugbear = "^22.4"
flake8-docstrings = "^1.6"
flake8-fixme = "^1.1"
flake8-type-checking = {version="^1.5", python="<4.0"}
isort = "^5.10"
mypy = "^0.950"
pep8-naming = "^0.12"
pre-commit = "^2.18"
pytest = "^7.1"
pytest-check = "^1.0"
pytest-cov = "^3.0"
pytest-randomly = "^3.12"
tox = "^3.25"

[tool.black]
line-length = 100

[tool.isort]
case_sensitive = true
known_first_party = "ppg_log,tests"
no_lines_before = "LOCALFOLDER"
order_by_type = false
profile = "black"
line_length = 100

[tool.mypy]
disallow_incomplete_defs = true
disallow_untyped_calls = true
disallow_untyped_decorators = true
disallow_untyped_defs = true
ignore_missing_imports = true
no_implicit_optional = true
show_error_codes = true
warn_redundant_casts = true
warn_return_any = true
warn_unused_configs = true
warn_unused_ignores = true

[build-system]
requires = ["poetry-core>=1.0.8"]
build-backend = "poetry.core.masonry.api"
